refactor: routes problems 관련 모듈화
Related to: #143
